requirements-dev.txt
black==24.2.0
isort==5.12.0
mypy==1.8.0
types-Pillow==10.2.0.20240213
types-tqdm==4.66.0.20240106
pydub-stubs==0.25.1.1